% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AT}
\alias{AT}
\title{Google stock prices.}
\format{
An xts object with 4048 rows and 2 variables:
\describe{
  \item{AMZN}{Adjusted price of Amazon stock, in US dollars}
  \item{GOOG.High}{Adjusted price of TESLA stock, in US dollars}
}
}
\source{
quantmod
}
\usage{
AT
}
\description{
An xts object containing the adjusted daily prices of Amazon and Tesla stock
}
\keyword{datasets}
