% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOOG}
\alias{GOOG}
\title{Google stock prices.}
\format{
A data frame with 3957 rows and 6 variables:
\describe{
  \item{GOOG.Open}{Open price, in US dollars}
  \item{GOOG.High}{Open price, in US dollars}
  \item{GOOG.Low}{Open price, in US dollars}
  \item{GOOG.Close}{Open price, in US dollars}
 \item{GOOG.Volume}{Open price, in US dollars}}
}
\source{
quantmod run
}
\usage{
GOOG
}
\description{
An xts object containing the daily prices of google stock
}
\keyword{datasets}
