% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS_aw.R
\name{HS_aw}
\alias{HS_aw}
\title{VaR and ES using Age-weighted HS}
\usage{
HS_aw(r, P = 1000, lam = 0.98, ci = 0.95)
}
\arguments{
\item{r}{a vector of returns}

\item{P}{portfolio value}

\item{lam}{lambda - the weighting parameter}

\item{ci}{confidence for VaR and ES}
}
\value{
a vector containing VaR and ES
}
\description{
VaR and ES using Age-weighted HS
}
\examples{
 x <- rnorm(100)
 HS_aw(x)
}
